Rename, restructure client subtopics; add Feign note
Each service type now has one consistently-named page on writing client
applications, and the dependency information and build file examples are
now in a new separate page.

This commit also adds a note on using Circuit Breaker Dashboard with
Feign, as the @HystrixCommand annotation cannot be applied directly to a
Feign client.