CS SI created in Apps Manager has no config source
A Config Server SI can be created but not configured in Apps Manager.
After creating a CS SI in Apps Manager, a user must update the SI using
the cf Command Line Interface tool and provide the SI with the URI of a
configuration source (Git repository). Move the note to this effect so
that it's at the beginning of the Apps Manager section (rather than at
the beginning of the topic) and mention at end of Apps Manager section
that the new SI must be updated using the cf CLI before it will be
usable.

[#126525717]